refactor: move charm_theme into internal/style
